+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 ### Unreleased
 
+### [3.0.0-alpha.6] - 2025-04-08
+
+- dep(eslint): upgraded to v9
+- dep(\*): bump versions to latest
+- feat(conf.d/http): added tls
+- feat(session): added JWT for session auth
+- feat(zone): removed location
+- feat(routes/zone_record): added, fixes #17
+- change(routes/users): result is always array
+- change(routes/ns): GET id is optional, result is always array
+- change(routes/zone): GET id is optional, result is always array
 
 ### [3.0.0-alpha.5] - 2024-03-06
 
@@ -31,3 +42,4 @@
 [3.0.0-alpha.3]: https://github.com/NicTool/api/releases/tag/3.0.0-alpha.3
 [3.0.0-alpha.4]: https://github.com/NicTool/api/releases/tag/3.0.0-alpha.4
 [3.0.0-alpha.5]: https://github.com/NicTool/api/releases/tag/3.0.0-alpha.5
+[3.0.0-alpha.6]: https://github.com/NicTool/api/releases/tag/3.0.0-alpha.6